requirements.txt
alabaster==0.7.12
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
atomicwrites==1.4.0
attrs==19.3.0
awesome-slugify==1.6.5
Babel==2.8.0
backcall==0.2.0
boto==2.49.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
colorama==0.4.3
coverage==5.2.1
decorator==4.4.2
defusedxml==0.6.0
Django==3.1
django-allauth==0.42.0
django-anymail==7.2.1
django-bootstrap-pagination==1.7.1
django-braces==1.14.0
django-coverage-plugin==1.8.0
django-crispy-forms==1.9.2
django-environ==0.4.5
django-extensions==3.0.4
django-filter==2.3.0
django-model-utils==4.0.0
django-redis==4.12.1
django-test-plus==1.4.0
djangorestframework==3.11.1
docutils==0.16
factory-boy==2.12.0
Faker==4.1.1
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4
idna==2.10
imagesize==1.2.0
iniconfig==1.0.1
ipython==7.17.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.2
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
oauthlib==3.1.0
packaging==20.4
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
prompt-toolkit==3.0.5
psycopg2==2.8.5
pudb==2019.2
py==1.9.0
pycparser==2.20
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
pytest==6.0.1
pytest-django==3.9.0
pytest-sugar==0.9.4
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.1
tqdm==4.48.2
traitlets==4.3.3
Unidecode==0.4.21
urllib3==1.25.10
urwid==2.1.1
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
zope.event==4.4
zope.interface==5.1.0
